pyproject.toml
[project]
name = "cdot65-scm"
version = "0.1.0"
description = "Ansible collection for Strata Cloud Manager (SCM) automation"
authors = [
    {name = "cdot65", email = "calvin@cdot.io"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"

[tool.poetry]
# This is an Ansible collection, not a standard Python package. Poetry is used for developer tooling only.
package-mode = false


[tool.poetry.dependencies]
ansible-core = "^2.17.0"
pan-scm-sdk = "^0.3.32"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
pytest-ansible = "^25.4.0"
pytest-xdist = "^3.6.1"
pytest-cov = "^6.1.1"
flake8 = "^6.1.0"
ruff = "^0.11.5"
mypy = "^1.15.0"
mock = "^5.2.0"
black = "^25.1.0"
isort = "^6.0.1"
python-dotenv = "^1.1.0"
ipython = "^9.2.0"
ansible-lint = "^25.4.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ansible.lint]
skip_list = ["no-handler"]

[tool.black]
line-length = 128
target-version = ['py310', 'py311', 'py312', 'py313']

[tool.ruff]
line-length = 128
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "D", "UP", "S", "BLE", "C4", "T20", "RET", "SIM", "ERA"]
ignore = ["E402", "E501", "D100", "D101", "D102", "D103", "D107", "D203", "D213"]

# Ignore some rules in tests files and Ansible code
[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S603", "S607", "T201", "ANN", "C408"]
"plugins/modules/*.py" = ["C408"]  # Ansible module convention uses dict() calls for module specs
"plugins/module_utils/*.py" = ["C408"]  # Ansible module utils also use dict() calls for argument specs

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.isort]
profile = "black"
line_length = 128
force_single_line = true
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "ANSIBLE_CORE", "ANSIBLE_CDOT65_SCM", "LOCALFOLDER"]
known_first_party = ["pan_scm_sdk"]
known_ansible_core = ["ansible"]
known_ansible_cdot65_scm = ["ansible_collections.cdot65.scm"]
from_first = false
# Skip Ansible modules as they have specific import requirements
skip_glob = ["plugins/modules/*.py"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

# NOTE:
# - All runtime dependencies for collection users (for Galaxy) must be listed in galaxy.yml, not here.
# - This file is for developer tooling, linting, and virtualenv management only.
# - Do NOT add collection metadata here for publishing to Galaxy; use galaxy.yml.
